change svg components to image
