+++ src/app/ui/index.tsx
@@ -1,12 +1,10 @@
 import "app/styles/index.css";
-
-// import { BoardPage } from "pages/board";
-import { WorkSpaceSettingsPage } from "src/pages/workspace/settings";
+import { AuthPage } from "src/pages/auth";
 
 export const App = () => {
   return (
     <>
-      <WorkSpaceSettingsPage />
+      <AuthPage />
     </>
   );
 };
