+++ src/pages/auth/index.ts
@@ -1,3 +1,3 @@
 export * from "./ui";
-export * as authModel from "./model";
+export * as authModel from "./model/model";
 export * as selectors from "./model/selectors";
