+++ src/pages/auth/model/selectors.ts
@@ -1,5 +1,5 @@
 import { useUnit } from "effector-react";
-import { $emailField, $isValid, actions } from ".";
+import { $emailField, $isValid, actions } from "./model";
 
 export const useEmailField = () =>
   useUnit([$emailField, actions.changedEmail, $isValid]);
